product, cathegory -- adding and editing

+++ Shop'sStorage/Models/Product.cs
@@ -11,7 +11,7 @@ namespace ShopSStorage.Models
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public int StorageAmount { get; set; }
-        public int DetailPrice { get; set; }
+        public decimal DetailPrice { get; set; }
         public virtual Cathegory Cathegory { get; set; }
         [Timestamp]
         public byte[] RowVersionBytes { get; set; }
